Fix 32 bit system test runs (#164)
* Fix 32 bit system test runs

Fixes #163

+++ s2/s2.go
@@ -110,7 +110,7 @@ func crc(b []byte) uint32 {
 
 // literalExtraSize returns the extra size of encoding n literals.
 // n should be >= 0 and <= math.MaxUint32.
-func literalExtraSize(n int) int {
+func literalExtraSize(n int64) int64 {
 	if n == 0 {
 		return 0
 	}
